Add files via upload
This is a data cleaning script in regard to a Nashville Housing dataset. In this script, we reformatted columns, populated null values, broke columns into multiple separate columns for organization and visual purposes, renamed columns, and deleted columns. This was all done in order to ensure that this set of data is more user-friendly.